--- entregas/oteroJorge/Ludoteca/src/Main.java
@@ -1,7 +1,0 @@
-import core.Aula;
-
-public class Main {
-    public static void main(String[] args) {
-        new Aula().aJugar();
-    }
-}
